# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.CommitCertRequest:
    description: 生成存证证书的请求体
    properties:
      evidence_level:
        description: 存证级别：internal/compliance/legal（或旧格式 L1/L2/L3）
        example: internal
        type: string
      trace_id:
        description: 追踪 ID
        example: trc_abc123
        type: string
    required:
    - trace_id
    type: object
...
  api.GenerateProofRequest:
    description: 生成最小披露证明的请求体
    properties:
      disclose_events:
        description: 要披露的事件索引（如 [0, 2]）
        items:
          type: integer
        type: array
      disclose_fields:
        description: 要披露的字段（如 ["output_hash"]）
        items:
          type: string
        type: array
    type: object
...
  cert.AnchorProof:
    properties:
      anchor_id:
        type: string
      anchor_timestamp:
        type: string
      anchor_type:
        description: local, worm, blockchain
        type: string
      blockchain:
        $ref: '#/definitions/cert.BlockchainProof'
      object_key:
        type: string
      storage_provider:
        type: string
    type: object
  cert.BlockchainProof:
    properties:
      block_height:
        type: integer
      chain_id:
        type: string
      contract_address:
        type: string
      tx_hash:
        type: string
    type: object
...
  cert.DisclosedEvent:
    properties:
      disclosed_fields:
        additionalProperties: true
        type: object
      event_hash:
        type: string
      event_index:
        type: integer
      event_type:
        type: string
    type: object
...
  cert.EventMerkleProof:
    properties:
      event_hash:
        type: string
      event_index:
        type: integer
      proof_path:
        items:
          $ref: '#/definitions/merkle.ProofNode'
        type: array
    type: object
...
  cert.MinimalDisclosureProof:
    properties:
      anchor_proof:
        allOf:
        - $ref: '#/definitions/cert.AnchorProof'
        description: 锚定证明
      cert_id:
        description: 证明目标
        type: string
      disclosed_events:
        description: 披露的事件
        items:
          $ref: '#/definitions/cert.DisclosedEvent'
        type: array
      merkle_proofs:
        description: Merkle证明
        items:
          $ref: '#/definitions/cert.EventMerkleProof'
        type: array
      root_hash:
        type: string
      schema_version:
        type: string
      time_proof:
        allOf:
        - $ref: '#/definitions/cert.TimeProof'
        description: 时间证明
      verification_instructions:
        allOf:
        - $ref: '#/definitions/cert.VerificationInstructions'
        description: 验证指引
    type: object
  cert.TimeProof:
    properties:
      proof_type:
        description: local, ntp, tsa
        type: string
      signature:
        type: string
      timestamp:
        type: string
      tsa_authority:
        type: string
      tsa_cert_chain:
        items:
          type: string
        type: array
      tsa_token:
        type: string
    type: object
  cert.VerificationInstructions:
    properties:
      verifier_url:
        type: string
      verify_command:
        type: string
    type: object
...
  merkle.ProofNode:
    properties:
      hash:
        type: string
      position:
        description: '"left" or "right"'
        type: string
    type: object
...
host: localhost:8006
info:
  contact:
    email: support@aitrace.cc
    name: AI-Trace Support
    url: https://aitrace.cc/support
  description: |-
    AI-Trace - Enterprise AI Decision Audit & Compliance Platform

    AI-Trace 提供零代码侵入的 AI 调用追踪、存证和审计能力。
    支持 OpenAI、Claude、Ollama 等多种模型，用户可透传自己的 API Key，服务不存储任何敏感信息。

    ## 特性
    - **零侵入集成**: 只需替换 API endpoint，无需修改代码
    - **API Key 透传**: 用户的 API Key 直接透传到上游，不存储
    - **自定义代理**: 支持用户指定自己的代理服务器
    - **多级存证**: internal(内部审计) / compliance(合规) / legal(法律效力)
    - **最小披露**: 基于 Merkle 树的选择性披露证明
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://aitrace.cc/terms
  title: AI-Trace API
  version: 0.2.0
paths:
...
  /certs/{cert_id}/prove:
    post:
      consumes:
      - application/json
      description: |-
        生成只包含必要信息的证明，支持选择性披露

        ## 使用场景
        - 向第三方证明 AI 决策过程，而不暴露完整对话内容
        - 选择性披露特定事件和字段
        - 生成可独立验证的 Merkle 证明
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 证书 ID
        in: path
        name: cert_id
        required: true
        type: string
      - description: 证明请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GenerateProofRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 最小披露证明
          schema:
            $ref: '#/definitions/cert.MinimalDisclosureProof'
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 证书不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 生成最小披露证明
      tags:
      - Certificates
  /certs/commit:
    post:
      consumes:
      - application/json
      description: |-
        为指定的追踪生成存证证书，支持多级存证

        ## 存证级别
        - **internal** (L1): Ed25519 签名，适合内部审计
        - **compliance** (L2): WORM 存储 + TSA 时间戳，适合合规要求
        - **legal** (L3): 区块链锚定，适合法律效力要求
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 存证请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CommitCertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 存证证书信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 追踪不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 生成存证证书
      tags:
      - Certificates
...
securityDefinitions:
  ApiKeyAuth:
    description: AI-Trace 平台 API Key（用于认证）
    in: header
    name: X-API-Key
    type: apiKey
  UpstreamApiKey:
    description: 用户的上游 API Key（OpenAI/Claude，透传不存储）
    in: header
    name: X-Upstream-API-Key
    type: apiKey
swagger: "2.0"